adjust the not found exception in delete function disease

+++ src/disease/disease.service.ts
@@ -98,17 +98,17 @@ export class DiseaseService {
   }
 
   async removeDiseaseRelation(id: number) {
-    const findId = await this.prisma.disease_symptoms.delete({
+    const findIdRelation = await this.findOneDiseaseRelation(id);
+
+    if (!findIdRelation) {
+      throw new NotFoundException('Id For Delete Is None!');
+    }
+
+    return this.prisma.disease_symptoms.delete({
       where: {
         disease_symptoms_id: id,
       },
     });
-
-    if (!findId) {
-      throw new NotFoundException('Id For Delete Is None!');
-    }
-
-    return findId;
   }
 
   async updateDiseaseRelation(
